Feat #15 EOF나기 전에 메뉴에서 파일 생성 및 첫번째 데이터 추가, 카운터 증가

+++ miniProject/src/consoleNotice/Menu.java
@@ -1,9 +1,12 @@
 package consoleNotice;
 
-import java.util.Scanner;
+import java.io.*;
+import java.util.*;
 
 public class Menu {
 	Scanner scanner = new Scanner(System.in);
+	File file = new File("/Users/sungin/Desktop/workspace/Eclipse/consoleBoard.db");
+	byte[] byteBuff = new byte[9999];
 	
 	public void showMenu() throws Exception {
 		BoardList boardList = new BoardList();
@@ -11,31 +14,60 @@ public class Menu {
 		Modify modify = new Modify();
 		Delete delete = new Delete();
 		SaveFile saveFile = new SaveFile();
-		
+
 		Scanner scanner = new Scanner(System.in);
 
-		System.out.println("----------------------------------------------------------------");
-		System.out.println("1.목록보기 | 2.상세보기 | 3.수정하기 | 4.삭제하기 | 5.파일저장 | 6.종료");
-		System.out.println("----------------------------------------------------------------");
-	
-		System.out.print("선택 : ");
-		String number = scanner.nextLine();
-	
-		switch(number) {
-		case "1": boardList.list(); break;
-		case "2": detail.detail(); break;
-		case "3": modify.modify(); break;
-		case "4": delete.delete(); break;
-		case "5": saveFile.saveFile(); break;
-		case "6": scanner.close(); return;
-		default : {
-			System.out.println("잘못된 입력입니다. 다시 입력하세요.");
-			showMenu();
+		boolean isExists = file.exists();
+		if(!isExists) file.createNewFile();
+		FileInputStream fis = new FileInputStream(file);
+		BufferedInputStream bis = new BufferedInputStream(fis);
+	    int nRLen = bis.read(byteBuff);
+	    if(nRLen < 0) {
+	    	System.out.println("------------------------");
+	    	System.out.println("처음 프로그램을 실행하셨군요!\n리스트 하나를 우선 등록해주세요!");
+	    	System.out.println("------------------------");
+	    	firstAddList();
+	    } else {
+	    	System.out.println("--------------------------------------------------------------------");
+			System.out.println("1.목록보기 | 2.상세보기 | 3.수정하기 | 4.삭제하기 | 5.파일 텍스트로 저장 | 6.종료");
+			System.out.println("--------------------------------------------------------------------");
+
+			System.out.print("선택 : ");
+			String number = scanner.nextLine();
+
+			switch(number) {
+			case "1": boardList.list(); break;
+			case "2": detail.detail(); break;
+			case "3": modify.modify(); break;
+			case "4": delete.delete(); break;
+			case "5": saveFile.saveFile(); break;
+			case "6": scanner.close(); return;
+			default : {
+				System.out.println("잘못된 입력입니다. 다시 입력하세요.");
+				showMenu();
+				}
 			}
 		}
 		
 		scanner.close();
 	}
+
+	public void firstAddList() throws Exception {
+		try {
+			List<Board> newList = new ArrayList<Board>();
+			System.out.println("타이틀, 내용, 작가를 순서대로 입력하세요.");
+			newList.add(new Board(1, scanner.nextLine(), scanner.nextLine(), scanner.nextLine(), new Date()));
+			FileOutputStream fos = new FileOutputStream("/Users/sungin/Desktop/workspace/Eclipse/consoleBoard.db");
+			ObjectOutputStream oos = new ObjectOutputStream(fos);
+			oos.writeObject(newList);
+			oos.flush();
+			oos.close();
+			showMenu();
+		} catch (Exception e) {
+			System.out.println("에러입니다. 에러 내용 : " + e);
+		}
+	}
+	
 	
 	public void listMenu() throws Exception {
 		BoardList boardList = new BoardList();
@@ -46,15 +78,24 @@ public class Menu {
 		case "2" : showMenu(); break;
 		default : {
 			System.out.println("잘못된 입력입니다. 다시 입력하세요.");
-			boardList.list();
+			listMenu();
 			break;
 		}
 		}
 	}
 	
 	public void detailMenu() throws Exception {
-		Detail detailList = new Detail();
+		Detail detail = new Detail();
 		System.out.print("선택 : ");
 		String number = scanner.nextLine();
+		switch (number) {
+		case "1" : detail.detail(); break;
+		case "2" : showMenu(); break;
+		default : {
+			System.out.println("잘못된 입력입니다. 다시 입력하세요.");
+			detailMenu();
+			break;
+		}
+		}
 	}
 }
